refactor: nuxi integration
