+++ src/nuxi/index.ts
@@ -1,4 +1,5 @@
-import { tryImport } from '../utils'
+import { window } from 'vscode'
+import { tryImport, isNuxiInstalled, newTerminal } from '../utils'
 
 export async function tryImportNuxi() {
     return (await tryImport('nuxi') || await tryImport('nuxi-edge')) as undefined | typeof import('nuxi-edge')
